renamed turn into host in gateway-core
