--- gateway/core/turn/__init__.py
@@ -1,1 +1,0 @@
-"""Running one gateway turn: the handler, its agents, its sink and its limits."""
