+++ gateway/tests/runtime/concurrency_limited_handler.py
@@ -1,6 +1,6 @@
 """Test-only capacity wrapper for arbitrary gateway callbacks.
 
-Production chat uses :class:`~gateway.core.turn.turn_handler.GatewayTurnHandler`
+Production chat uses :class:`~gateway.core.host.turn_handler.GatewayTurnHandler`
 with ``gate=``. This helper stays under ``gateway/tests/`` so it cannot be
 mistaken for a second production turn-handler class (Wave C4 quarantine).
 """
@@ -10,8 +10,8 @@ from __future__ import annotations
 import logging
 
 from core.agent_harness.session import SessionCore
+from gateway.core.host.concurrency import TurnConcurrencyGate
 from gateway.core.transport_api import GatewayAgentCallback, GatewaySink
-from gateway.core.turn.concurrency import TurnConcurrencyGate
 
 
 class ConcurrencyLimitedTurnHandler:
