+++ gateway/transports/discord/settings.py
@@ -11,8 +11,8 @@ from pydantic_settings import BaseSettings, NoDecode, SettingsConfigDict
 
 from config.constants.discord import DISCORD_ALLOW_OPEN_GUILD_ENV, DISCORD_ALLOWED_USERS_ENV
 from config.strict_config import StrictConfigModel
+from gateway.core.host.concurrency import turn_limit_for_profile
 from gateway.core.runtime.errors import GatewayConfigurationError
-from gateway.core.turn.concurrency import turn_limit_for_profile
 from integrations.messaging_security import MessagingIdentityPolicy, MessagingPlatform
 from integrations.store import get_integration
