+++ tests/core/agent/orchestration/test_cross_surface_parity.py
@@ -19,7 +19,7 @@ import pytest
 
 import surfaces.interactive_shell.runtime.slash_adapter as slash_adapter
 from core.agent_harness.tools.action_tools import get_action_tool
-from gateway.core.turn.turn_handler import GatewayTurnHandler
+from gateway.core.host.turn_handler import GatewayTurnHandler
 from tests.core.agent.orchestration.cross_surface_parity_harness import (
     ALL_SURFACES,
     PARITY_ANSWER,
